+++ tree/phylosupertreeplen.cpp
@@ -250,7 +250,7 @@ void PhyloSuperTreePlen::linkTrees() {
 double PhyloSuperTreePlen::optimizeAllBranches(int my_iterations, double tolerance, int maxNRStep) {
 	//initPartitionInfo(); // OLGA: not needed here
 	//cout<<"Optimizing all branches"<<endl;
-	for(int part = 0; part < size(); part++){
+	for(size_t part = 0; part < size(); part++){
 		part_info[part].cur_score = 0.0;
 	}
 
@@ -1500,7 +1500,7 @@ void PhyloSuperTreePlen::mapBranchLen()
 	int i;
 	getBranches(nodes1, nodes2);
 	double *checkVAL = new double[branchNum];
-	for(int part = 0; part < size(); part++){
+	for(size_t part = 0; part < size(); part++){
 		memset(checkVAL,0,at(part)->branchNum*sizeof(double));
 		for (i = 0; i < nodes1.size(); i++){
 			if(((SuperNeighbor*)nodes1[i]->findNeighbor(nodes2[i]))->link_neighbors[part])
@@ -1509,7 +1509,7 @@ void PhyloSuperTreePlen::mapBranchLen()
 		}
 		NodeVector nodes1_sub, nodes2_sub;
 		at(part)->getBranches(nodes1_sub, nodes2_sub);
-		for(int j = 0; j<nodes1_sub.size();j++){
+		for(size_t j = 0; j<nodes1_sub.size();j++){
 			nodes1_sub[j]->findNeighbor(nodes2_sub[j])->length = checkVAL[nodes1_sub[j]->findNeighbor(nodes2_sub[j])->id];
 			nodes2_sub[j]->findNeighbor(nodes1_sub[j])->length = checkVAL[nodes1_sub[j]->findNeighbor(nodes2_sub[j])->id];
 		}
@@ -1531,7 +1531,7 @@ void PhyloSuperTreePlen::mapBranchLen(int part)
     }
     NodeVector nodes1_sub, nodes2_sub;
     at(part)->getBranches(nodes1_sub, nodes2_sub);
-    for(int j = 0; j<nodes1_sub.size();j++){
+    for(size_t j = 0; j<nodes1_sub.size();j++){
         nodes1_sub[j]->findNeighbor(nodes2_sub[j])->length = checkVAL[nodes1_sub[j]->findNeighbor(nodes2_sub[j])->id];
         nodes2_sub[j]->findNeighbor(nodes1_sub[j])->length = checkVAL[nodes1_sub[j]->findNeighbor(nodes2_sub[j])->id];
     }
